ga1/q19/vector-similarity-api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.54.3
numpy==1.26.3
python-dotenv==1.0.0
httpx==0.27.0